>[project:chezmoi.sh]: Rename pulumi resources dir src/ to stack/
The src/ dir holds the Pulumi stack's resource definitions. Renaming
to stack/ pairs with toolbox/ for a clear semantic layout: stack/
for resources, toolbox/ for tooling, root for entry and config.

No behavior change; live stack state is unaffected (keyed by URNs in
the S3 backend, not file paths).
