+++ projects/chezmoi.sh/src/infrastructure/pulumi/.mocharc.json
@@ -1,6 +1,6 @@
 {
 	"require": ["ts-node/register"],
-	"spec": ["src/**/*.test.ts", "toolbox/**/*.test.ts"],
+	"spec": ["stack/**/*.test.ts", "toolbox/**/*.test.ts"],
 	"timeout": 10000,
 	"node-option": ["disable-warning=MODULE_TYPELESS_PACKAGE_JSON"]
 }
