+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/observability.ts
