+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/cloudsync.ts
