+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/shares.ts
