+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/users/firesticktv.ts
