+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/zpools/README.md
