+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/zpools/zp1cs01.ts
