+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/truenas/zpools/zp1hs01.ts
