+++ projects/chezmoi.sh/src/infrastructure/pulumi/stack/zot-registry.ts
