+++ projects/chezmoi.sh/src/infrastructure/pulumi/tsconfig.json
@@ -3,5 +3,5 @@
 	"compilerOptions": {
 		"outDir": "bin"
 	},
-	"include": ["*.ts", "src/**/*.ts", "toolbox/**/*.ts"]
+	"include": ["*.ts", "stack/**/*.ts", "toolbox/**/*.ts"]
 }
