+++ projects/chezmoi.sh/src/infrastructure/pulumi/tsconfig.test.json
@@ -4,5 +4,5 @@
 		"noEmit": true,
 		"types": ["mocha", "node", "chai"]
 	},
-	"include": ["*.ts", "src/**/*.ts", "toolbox/**/*.ts"]
+	"include": ["*.ts", "stack/**/*.ts", "toolbox/**/*.ts"]
 }
